requirements.txt
bleach==1.5.0
cycler==0.10.0
enum34==1.1.6
futures==3.1.1
html5lib==0.9999999
Markdown==2.6.9
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pandas==0.20.3
Pillow==4.2.1
pip-autoremove==0.9.0
pipdeptree==0.10.1
protobuf==3.5.0.post1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc1
tflearn==0.3.2
tqdm==4.19.4
Werkzeug==0.12.2
